Move utility functions and libraries (print, assets, HTTP and VFS) out of ovms_duktape
