+++ vehicle/OVMS.V3/components/ovms_script/srcduk/ovms_duktape.cpp
@@ -407,7 +407,7 @@ void DukOvmsFatalHandler(void *udata, const char *msg)
   abort();
   }
 
-void DukOvmsErrorHandler(duk_context *ctx, duk_idx_t err_idx, OvmsWriter *writer=NULL, const char *filename=NULL)
+void DukOvmsErrorHandler(duk_context *ctx, duk_idx_t err_idx, OvmsWriter *writer, const char *filename)
   {
   const char *error;
   int linenumber = 0;
@@ -514,67 +514,6 @@ static duk_ret_t DukOvmsLoadModule(duk_context *ctx)
     }
 
   return 1;
-  }
-
-static void DukGetCallInfo(duk_context *ctx, std::string *filename, int *linenumber, std::string *function)
-  {
-  duk_require_stack(ctx, 3);
-  for (int i=-2; ; i--)
-    {
-    *filename = "";
-    *function = "";
-    duk_inspect_callstack_entry(ctx, i);
-    duk_get_prop_string(ctx, -1, "lineNumber");
-    *linenumber = duk_get_number_default(ctx, -1, 0);
-    duk_pop(ctx);
-    if (duk_get_prop_string(ctx, -1, "function"))
-      {
-      duk_get_prop_string(ctx, -1, "fileName");
-      *filename = duk_get_string_default(ctx, -1, "");
-      duk_pop(ctx);
-      duk_get_prop_string(ctx, -1, "name");
-      *function = duk_get_string_default(ctx, -1, "");
-      duk_pop(ctx);
-      }
-    duk_pop(ctx);
-    // skip internal modules:
-    if (!startsWith(*filename, "int/"))
-      break;
-    }
-  }
-
-static duk_ret_t DukOvmsPrint(duk_context *ctx)
-  {
-  const char *output = duk_safe_to_string(ctx,0);
-
-  if (output != NULL)
-    {
-    if (duktapewriter != NULL)
-      {
-      duktapewriter->printf("%s",output);
-      }
-    else
-      {
-      std::string filename, function;
-      int linenumber = 0;
-      DukGetCallInfo(ctx, &filename, &linenumber, &function);
-      if (function == "eval")
-        ESP_LOGI(TAG, "[%s:%d] %s", filename.c_str(), linenumber, output);
-      else
-        ESP_LOGI(TAG, "[%s:%d:%s] %s", filename.c_str(), linenumber, function.c_str(), output);
-      }
-    }
-  return 0;
-  }
-
-static duk_ret_t DukOvmsAssert(duk_context *ctx)
-  {
-  if (duk_to_boolean(ctx, 0))
-    {
-    return 0;
-    }
-  duk_error(ctx, DUK_ERR_ERROR, "assertion failed: %s", duk_safe_to_string(ctx, 1));
-  return 0;
   }
 
 ////////////////////////////////////////////////////////////////////////////////
@@ -873,898 +812,6 @@ duk_ret_t DuktapeObject::CallMethod(duk_context *ctx, const char* method, void* 
   }
 
 ////////////////////////////////////////////////////////////////////////////////
-//DuktapeHTTPRequest
-
-DuktapeHTTPRequest::DuktapeHTTPRequest(duk_context *ctx, int obj_idx)
-  : DuktapeObject(ctx, obj_idx)
-  {
-  // get args:
-  duk_require_stack(ctx, 5);
-  if (duk_get_prop_string(ctx, 0, "url"))
-    m_url = duk_to_string(ctx, -1);
-  duk_pop(ctx);
-  if (duk_get_prop_string(ctx, 0, "post"))
-    {
-    m_ispost = true;
-    m_post = duk_to_string(ctx, -1);
-    }
-  duk_pop(ctx);
-  if (duk_get_prop_string(ctx, 0, "timeout"))
-    m_timeout = duk_to_int(ctx, -1);
-  duk_pop(ctx);
-  if (duk_get_prop_string(ctx, 0, "binary"))
-    m_binary = duk_to_boolean(ctx, -1);
-  duk_pop(ctx);
-
-  if (m_url.empty())
-    {
-    m_error = "missing argument: url";
-    CallMethod(ctx, "fail");
-    return;
-    }
-
-  // …request headers:
-  extram::string key, val;
-  bool have_useragent = false, have_contenttype = false;
-
-  duk_get_prop_string(ctx, 0, "headers");
-  if (duk_is_array(ctx, -1))
-    {
-    for (int i=0; true; i++)
-      {
-      if (!duk_get_prop_index(ctx, -1, i))
-        {
-        // array end
-        duk_pop(ctx);
-        break;
-        }
-      if (duk_is_object(ctx, -1))
-        {
-        duk_enum(ctx, -1, 0);
-        while (duk_next(ctx, -1, true))
-          {
-          key = duk_to_string(ctx, -2);
-          val = duk_to_string(ctx, -1);
-          duk_pop_2(ctx);
-          m_headers.append(key);
-          m_headers.append(": ");
-          m_headers.append(val);
-          m_headers.append("\r\n");
-          if (key == "User-Agent")
-            have_useragent = true;
-          else if (key == "Content-Type")
-            have_contenttype = true;
-          }
-        duk_pop(ctx); // enum
-        }
-      duk_pop(ctx); // array element
-      }
-    }
-  duk_pop(ctx); // [array]
-
-  // add defaults:
-  if (!have_useragent)
-    {
-    m_headers.append("User-Agent: ");
-    m_headers.append(get_user_agent().c_str());
-    m_headers.append("\r\n");
-    }
-  if (m_ispost && !have_contenttype)
-    {
-    m_headers.append("Content-Type: application/x-www-form-urlencoded\r\n");
-    }
-
-  // check network:
-  if (!MyNetManager.MongooseRunning() || !MyNetManager.m_connected_any)
-    {
-    m_error = "network unavailable";
-    CallMethod(ctx, "fail");
-    return;
-    }
-
-  // start initial request:
-  if (StartRequest(ctx))
-    {
-    // running, prevent deletion & GC:
-    Ref();
-    Register(ctx);
-    ESP_LOGD(TAG, "DuktapeHTTPRequest: started '%s'", m_url.c_str());
-    }
-  }
-
-bool DuktapeHTTPRequest::StartRequest(duk_context *ctx /*=NULL*/)
-  {
-  // create connection:
-  m_mgconn = NULL;
-  struct mg_mgr* mgr = MyNetManager.GetMongooseMgr();
-  struct mg_connect_opts opts = {};
-  opts.user_data = this;
-  const char* err;
-  opts.error_string = &err;
-
-  if (startsWith(m_url, "https://"))
-    {
-    #if MG_ENABLE_SSL
-      opts.ssl_ca_cert = MyOvmsTLS.GetTrustedList();
-    #else
-      m_error = "SSL support disabled";
-      ESP_LOGD(TAG, "DuktapeHTTPRequest: connect to '%s' failed: %s", m_url.c_str(), m_error.c_str());
-      CallMethod(ctx, "fail");
-      return false;
-    #endif
-    }
-
-  m_mgconn = mg_connect_http_opt(mgr, MongooseCallbackEntry, opts,
-    m_url.c_str(), m_headers.c_str(), m_ispost ? m_post.c_str() : NULL);
-
-  if (!m_mgconn)
-    {
-    ESP_LOGD(TAG, "DuktapeHTTPRequest: connect to '%s' failed: %s", m_url.c_str(), err);
-    m_error = (err && *err) ? err : "unknown";
-    CallMethod(ctx, "fail");
-    return false;
-    }
-
-  // connection created:
-  if (m_timeout > 0)
-    mg_set_timer(m_mgconn, mg_time() + (double)m_timeout / 1000);
-  return true;
-  }
-
-DuktapeHTTPRequest::~DuktapeHTTPRequest()
-  {
-  // ESP_LOGD(TAG, "~DuktapeHTTPRequest");
-  if (m_mgconn)
-    {
-    m_mgconn->user_data = NULL;
-    m_mgconn->flags |= MG_F_CLOSE_IMMEDIATELY;
-    m_mgconn = NULL;
-    }
-  }
-
-duk_ret_t DuktapeHTTPRequest::Create(duk_context *ctx)
-  {
-  // var request = HTTP.Request({ args })
-  DuktapeHTTPRequest* request = new DuktapeHTTPRequest(ctx, 0);
-  request->Push(ctx);
-  return 1;
-  }
-
-void DuktapeHTTPRequest::MongooseCallbackEntry(struct mg_connection *nc, int ev, void *ev_data)
-  {
-  DuktapeHTTPRequest* me = (DuktapeHTTPRequest*)nc->user_data;
-  if (me) me->MongooseCallback(nc, ev, ev_data);
-  }
-
-void DuktapeHTTPRequest::MongooseCallback(struct mg_connection *nc, int ev, void *ev_data)
-  {
-  OvmsRecMutexLock lock(&m_mutex);
-  if (nc != m_mgconn) return; // ignore events of previous connections
-  switch (ev)
-    {
-    case MG_EV_CONNECT:
-      {
-      int err = *(int*) ev_data;
-      const char* errdesc = strerror(err);
-      ESP_LOGD(TAG, "DuktapeHTTPRequest: MG_EV_CONNECT err=%d/%s", err, errdesc);
-      if (err)
-        {
-        #if MG_ENABLE_SSL
-        if (err == MG_SSL_ERROR)
-          m_error = "SSL error";
-        else
-        #endif
-          m_error = (errdesc && *errdesc) ? errdesc : "unknown";
-        RequestCallback("fail");
-        nc->flags |= MG_F_CLOSE_IMMEDIATELY;
-        }
-      }
-      break;
-
-    case MG_EV_HTTP_REPLY:
-      {
-      // response is complete, store:
-      http_message *hm = (http_message*) ev_data;
-      ESP_LOGD(TAG, "DuktapeHTTPRequest: MG_EV_HTTP_REPLY status=%d bodylen=%d", hm->resp_code, hm->body.len);
-      m_response_status = hm->resp_code;
-      m_response_statusmsg.assign(hm->resp_status_msg.p, hm->resp_status_msg.len);
-      m_response_body.assign(hm->body.p, hm->body.len);
-      extram::string key, val, location;
-      for (int i = 0; i < MG_MAX_HTTP_HEADERS && hm->header_names[i].len > 0; i++)
-        {
-        key.assign(hm->header_names[i].p, hm->header_names[i].len);
-        val.assign(hm->header_values[i].p, hm->header_values[i].len);
-        m_response_headers.push_back(std::make_pair(key, val));
-        if (key == "Location") location = val;
-        }
-
-      // follow redirect?
-      if (m_response_status == 301 || m_response_status == 302)
-        {
-        if (location.empty())
-          {
-          m_error = "redirect without location";
-          RequestCallback("fail");
-          }
-        else if (++m_redirectcnt > 5)
-          {
-          m_error = "too many redirects";
-          RequestCallback("fail");
-          }
-        else
-          {
-          ESP_LOGD(TAG, "DuktapeHTTPRequest: redirect code=%d to '%s'", m_response_status, location.c_str());
-          m_url = location;
-          m_response_status = 0;
-          m_response_statusmsg.clear();
-          m_response_body.clear();
-          m_response_headers.clear();
-          if (!StartRequest(NULL))
-            RequestCallback("fail");
-          }
-        }
-      else
-        {
-        RequestCallback("done");
-        }
-      // in any case, this connection is done:
-      nc->flags |= MG_F_CLOSE_IMMEDIATELY;
-      }
-      break;
-
-    case MG_EV_TIMER:
-      {
-      ESP_LOGD(TAG, "DuktapeHTTPRequest: MG_EV_TIMER");
-      m_error = "timeout";
-      RequestCallback("fail");
-      nc->flags |= MG_F_CLOSE_IMMEDIATELY;
-      }
-      break;
-
-    case MG_EV_CLOSE:
-      {
-      if (m_response_status == 0 && m_error.empty())
-        {
-        ESP_LOGD(TAG, "DuktapeHTTPRequest: MG_EV_CLOSE: abort");
-        m_error = "abort";
-        RequestCallback("fail");
-        }
-      else
-        {
-        ESP_LOGD(TAG, "DuktapeHTTPRequest: MG_EV_CLOSE status=%d", m_response_status);
-        }
-      // Mongoose part done:
-      nc->user_data = NULL;
-      m_mgconn = NULL;
-      Unref();
-      }
-      break;
-
-    default:
-      break;
-    }
-  }
-
-duk_ret_t DuktapeHTTPRequest::CallMethod(duk_context *ctx, const char* method, void* data /*=NULL*/)
-  {
-  if (!ctx)
-    {
-    RequestCallback(method, data);
-    return 0;
-    }
-  OvmsRecMutexLock lock(&m_mutex);
-  if (!IsCoupled()) return 0;
-
-  duk_require_stack(ctx, 7);
-  int entry_top = duk_get_top(ctx);
-
-  bool deregister = false;
-
-  while (method)
-    {
-    const char* followup_method = NULL;
-
-    // check method:
-    int obj_idx = Push(ctx);
-    duk_get_prop_string(ctx, obj_idx, method);
-    bool callable = duk_is_callable(ctx, -1);
-    duk_pop(ctx);
-    if (callable) duk_push_string(ctx, method);
-    int nargs = 0;
-
-    // update request.url, set request.redirectCount:
-    duk_push_string(ctx, m_url.c_str());
-    duk_put_prop_string(ctx, obj_idx, "url");
-    duk_push_int(ctx, m_redirectcnt);
-    duk_put_prop_string(ctx, obj_idx, "redirectCount");
-
-    // create results & method arguments:
-    if (strcmp(method, "done") == 0)
-      {
-      // done(response):
-      followup_method = "always";
-      ESP_LOGD(TAG, "DuktapeHTTPRequest: done status=%d bodylen=%d url='%s'",
-        m_response_status, m_response_body.size(), m_url.c_str());
-
-      // clear request.error:
-      duk_push_string(ctx, "");
-      duk_put_prop_string(ctx, obj_idx, "error");
-
-      // create response object:
-      duk_push_object(ctx);
-      duk_push_int(ctx, m_response_status);
-      duk_put_prop_string(ctx, -2, "statusCode");
-      duk_push_string(ctx, m_response_statusmsg.c_str());
-      duk_put_prop_string(ctx, -2, "statusText");
-
-      // …body:
-      if (m_binary)
-        {
-        void* p = duk_push_fixed_buffer(ctx, m_response_body.size());
-        memcpy(p, m_response_body.data(), m_response_body.size());
-        duk_put_prop_string(ctx, -2, "data");
-        }
-      else
-        {
-        duk_push_lstring(ctx, m_response_body.data(), m_response_body.size());
-        duk_put_prop_string(ctx, -2, "body");
-        }
-      m_response_body.clear();
-      m_response_body.shrink_to_fit();
-
-      // …response headers:
-      duk_push_array(ctx);
-      int i = 0;
-      for (auto it = m_response_headers.begin(); it != m_response_headers.end(); it++, i++)
-        {
-        duk_push_object(ctx);
-        duk_push_string(ctx, it->second.c_str());
-        duk_put_prop_string(ctx, -2, it->first.c_str());
-        duk_put_prop_index(ctx, -2, i);
-        }
-      duk_put_prop_string(ctx, -2, "headers");
-      m_response_headers.clear();
-
-      duk_dup(ctx, -1);
-      duk_put_prop_string(ctx, obj_idx, "response");
-      nargs++;
-      }
-    else if (strcmp(method, "fail") == 0)
-      {
-      // fail(error):
-      followup_method = "always";
-      ESP_LOGD(TAG, "DuktapeHTTPRequest: failed error='%s' url='%s'",
-        m_error.c_str(), m_url.c_str());
-
-      // set request.error:
-      duk_push_string(ctx, m_error.c_str());
-      duk_dup(ctx, -1);
-      duk_put_prop_string(ctx, obj_idx, "error");
-      nargs++;
-      }
-    else if (strcmp(method, "always") == 0)
-      {
-      // always():
-      deregister = true;
-      }
-
-    // call method:
-    if (callable)
-      {
-      ESP_LOGD(TAG, "DuktapeHTTPRequest: calling method '%s' nargs=%d", method, nargs);
-      if (duk_pcall_prop(ctx, obj_idx, nargs) != 0)
-        {
-        DukOvmsErrorHandler(ctx, -1);
-        }
-      }
-
-    // clear stack:
-    duk_pop_n(ctx, duk_get_top(ctx) - entry_top);
-
-    // followup call:
-    method = followup_method;
-    } // while (method)
-
-  // allow GC:
-  if (deregister) Deregister(ctx);
-  return 0;
-  }
-
-////////////////////////////////////////////////////////////////////////////////
-// DuktapeVFSLoad: load a file asynchronously
-//
-// This is following a high-level approach, limitation is the file contents needs to fit in RAM
-// twice, as the read buffer is converted into a JS string/buffer.
-// Partial loading / block operations may be added later as needed, e.g. based on arguments.
-//
-// Notes:
-// - Conditional synchronous operation doesn't make sense; async overhead is minimal
-//   and the stat() on "/sd" already needs at least 45 ms.
-// - On /store a Load() takes ~ 25 ms base +  5 ms per KB.
-// - On /sd    a Load() takes ~ 60 ms base + 30 ms per KB.
-//
-// Javascript API:
-//   var request = VFS.Load({
-//     path: "…",
-//     [binary: false,]
-//     done: function(data) {},      // data: string|u8buf
-//     fail: function(error) {},
-//     always: function() {},
-//   });
-
-class DuktapeVFSLoad : public DuktapeObject
-  {
-  public:
-    DuktapeVFSLoad(duk_context *ctx, int obj_idx);
-    ~DuktapeVFSLoad() { /*ESP_LOGD(TAG, "~DuktapeVFSLoad");*/ }
-    static duk_ret_t Create(duk_context *ctx);
-
-  public:
-    duk_ret_t CallMethod(duk_context *ctx, const char* method, void* data=NULL);
-
-  protected:
-    static void LoadTask(void *param);
-    void Load();
-
-  protected:
-    std::ifstream m_file;
-    std::string m_path;
-    struct stat m_stat;
-    extram::string m_data;
-    std::string m_error;
-    bool m_binary = false;
-  };
-
-duk_ret_t DuktapeVFSLoad::Create(duk_context *ctx)
-  {
-  // var request = VFS.Load({ args })
-  DuktapeVFSLoad *request = new DuktapeVFSLoad(ctx, 0);
-  request->Push(ctx);
-  return 1;
-  }
-
-DuktapeVFSLoad::DuktapeVFSLoad(duk_context *ctx, int obj_idx)
-  : DuktapeObject(ctx, obj_idx)
-  {
-  // get args:
-  duk_require_stack(ctx, 5);
-  if (duk_get_prop_string(ctx, 0, "path"))
-    m_path = duk_to_string(ctx, -1);
-  duk_pop(ctx);
-  if (duk_get_prop_string(ctx, 0, "binary"))
-    m_binary = duk_to_boolean(ctx, -1);
-  duk_pop(ctx);
-
-  // check path:
-  if (m_path.empty())
-    {
-    m_error = "no path";
-    CallMethod(ctx, "fail");
-    return;
-    }
-  if (MyConfig.ProtectedPath(m_path))
-    {
-    m_error = "protected path";
-    CallMethod(ctx, "fail");
-    return;
-    }
-
-  // start loader:
-  Ref();
-  Register(ctx);
-  if (xTaskCreatePinnedToCore(LoadTask, "DuktapeVFSLoad", 5*512, this,
-      CONFIG_OVMS_SC_JAVASCRIPT_DUKTAPE_PRIORITY-1, NULL, CORE(1)) != pdPASS)
-    {
-    Deregister(ctx);
-    Unref();
-    m_error = "cannot start task";
-    CallMethod(ctx, "fail");
-    return;
-    }
-  //ESP_LOGD(TAG, "DuktapeVFSLoad('%s'): started", m_path.c_str());
-  }
-
-void DuktapeVFSLoad::LoadTask(void *param)
-  {
-  DuktapeVFSLoad *me = (DuktapeVFSLoad*)param;
-  me->Load();
-  me->Unref();
-  vTaskDelete(NULL);
-  }
-
-void DuktapeVFSLoad::Load()
-  {
-  // check file & size:
-  if (stat(m_path.c_str(), &m_stat) != 0)
-    {
-    m_error = "file not found";
-    RequestCallback("fail");
-    return;
-    }
-  if (S_ISDIR(m_stat.st_mode))
-    {
-    m_error = "not a file";
-    RequestCallback("fail");
-    return;
-    }
-  if (m_stat.st_size > heap_caps_get_largest_free_block(MALLOC_CAP_SPIRAM))
-    {
-    m_error = "file too large";
-    RequestCallback("fail");
-    return;
-    }
-  if (m_stat.st_size == 0)
-    {
-    m_error = "";
-    RequestCallback("done");
-    return;
-    }
-
-  // load data:
-  m_file.open(m_path, std::ios::in | std::ios::binary);
-  if (m_file.fail())
-    {
-    m_error = "cannot open file";
-    RequestCallback("fail");
-    return;
-    }
-  m_data.resize(m_stat.st_size, '\0');
-  m_file.read(&m_data[0], m_stat.st_size);
-  bool fail = m_file.fail();
-  m_file.close();
-
-  if (fail)
-    {
-    m_error = "read error";
-    RequestCallback("fail");
-    }
-  else
-    {
-    m_error = "";
-    RequestCallback("done");
-    }
-  }
-
-duk_ret_t DuktapeVFSLoad::CallMethod(duk_context *ctx, const char* method, void* data /*=NULL*/)
-  {
-  if (!ctx)
-    {
-    RequestCallback(method, data);
-    return 0;
-    }
-  OvmsRecMutexLock lock(&m_mutex);
-  if (!IsCoupled()) return 0;
-
-  duk_require_stack(ctx, 3);
-  int entry_top = duk_get_top(ctx);
-
-  bool deregister = false;
-
-  while (method)
-    {
-    const char* followup_method = NULL;
-
-    // check method:
-    int obj_idx = Push(ctx);
-    duk_get_prop_string(ctx, obj_idx, method);
-    bool callable = duk_is_callable(ctx, -1);
-    duk_pop(ctx);
-    if (callable) duk_push_string(ctx, method);
-    int nargs = 0;
-
-    // create results & method arguments:
-    if (strcmp(method, "done") == 0)
-      {
-      // done(data):
-      followup_method = "always";
-      //ESP_LOGD(TAG, "DuktapeVFSLoad('%s'): done size=%d", m_path.c_str(), m_data.size());
-
-      // clear request.error:
-      duk_push_string(ctx, "");
-      duk_put_prop_string(ctx, obj_idx, "error");
-
-      // set data:
-      if (m_binary)
-        {
-        void* p = duk_push_fixed_buffer(ctx, m_data.size());
-        memcpy(p, m_data.data(), m_data.size());
-        duk_dup(ctx, -1);
-        duk_put_prop_string(ctx, obj_idx, "data");
-        }
-      else
-        {
-        duk_push_lstring(ctx, m_data.data(), m_data.size());
-        duk_dup(ctx, -1);
-        duk_put_prop_string(ctx, obj_idx, "data");
-        }
-      m_data.clear();
-      m_data.shrink_to_fit();
-      nargs++;
-      }
-    else if (strcmp(method, "fail") == 0)
-      {
-      // fail(error):
-      followup_method = "always";
-      //ESP_LOGD(TAG, "DuktapeVFSLoad('%s'): failed error='%s'", m_path.c_str(), m_error.c_str());
-
-      // set request.error:
-      duk_push_string(ctx, m_error.c_str());
-      duk_dup(ctx, -1);
-      duk_put_prop_string(ctx, obj_idx, "error");
-      nargs++;
-      }
-    else if (strcmp(method, "always") == 0)
-      {
-      // always():
-      deregister = true;
-      }
-
-    // call method:
-    if (callable)
-      {
-      //ESP_LOGD(TAG, "DuktapeVFSLoad: calling method '%s' nargs=%d", method, nargs);
-      if (duk_pcall_prop(ctx, obj_idx, nargs) != 0)
-        {
-        DukOvmsErrorHandler(ctx, -1);
-        }
-      }
-
-    // clear stack:
-    duk_pop_n(ctx, duk_get_top(ctx) - entry_top);
-
-    // followup call:
-    method = followup_method;
-    } // while (method)
-
-  // allow GC:
-  if (deregister) Deregister(ctx);
-  return 0;
-  }
-
-////////////////////////////////////////////////////////////////////////////////
-// DuktapeVFSSave: save a file asynchronously
-//
-// This is following a high-level approach, limitation is the file contents needs to fit in RAM
-// twice, as the JS buffer is copied for the saver.
-// Partial saving / block operations may be added later as needed, e.g. based on arguments.
-//
-// Javascript API:
-//   var request = VFS.Save({
-//     data: string|u8buf
-//     path: "…",                 // missing directories will be created automatically
-//     [append: false,]
-//     done: function() {},
-//     fail: function(error) {},
-//     always: function() {},
-//   });
-
-class DuktapeVFSSave : public DuktapeObject
-  {
-  public:
-    DuktapeVFSSave(duk_context *ctx, int obj_idx);
-    ~DuktapeVFSSave() { /*ESP_LOGD(TAG, "~DuktapeVFSSave");*/ }
-    static duk_ret_t Create(duk_context *ctx);
-
-  public:
-    duk_ret_t CallMethod(duk_context *ctx, const char* method, void* data=NULL);
-
-  protected:
-    static void SaveTask(void *param);
-    void Save();
-
-  protected:
-    std::ofstream m_file;
-    std::string m_path;
-    extram::string m_data;
-    std::string m_error;
-    bool m_append = false;
-  };
-
-duk_ret_t DuktapeVFSSave::Create(duk_context *ctx)
-  {
-  // var request = VFS.Save({ args })
-  DuktapeVFSSave *request = new DuktapeVFSSave(ctx, 0);
-  request->Push(ctx);
-  return 1;
-  }
-
-DuktapeVFSSave::DuktapeVFSSave(duk_context *ctx, int obj_idx)
-  : DuktapeObject(ctx, obj_idx)
-  {
-  // get args:
-  duk_require_stack(ctx, 5);
-  if (duk_get_prop_string(ctx, 0, "path"))
-    m_path = duk_to_string(ctx, -1);
-  duk_pop(ctx);
-  if (duk_get_prop_string(ctx, 0, "append"))
-    m_append = duk_to_boolean(ctx, -1);
-  duk_pop(ctx);
-
-  if (duk_get_prop_string(ctx, 0, "data"))
-    {
-    if (duk_is_buffer_data(ctx, -1))
-      {
-      size_t size;
-      void *data = duk_get_buffer_data(ctx, -1, &size);
-      m_data.resize(size, '\0');
-      memcpy(&m_data[0], data, size);
-      }
-    else
-      {
-      m_data = duk_to_string(ctx, -1);
-      }
-    duk_pop(ctx);
-    }
-  else
-    {
-    duk_pop(ctx);
-    m_error = "no data";
-    CallMethod(ctx, "fail");
-    return;
-    }
-
-  // check path:
-  if (m_path.empty())
-    {
-    m_error = "no path";
-    CallMethod(ctx, "fail");
-    return;
-    }
-  if (MyConfig.ProtectedPath(m_path))
-    {
-    m_error = "protected path";
-    CallMethod(ctx, "fail");
-    return;
-    }
-
-  // start saver:
-  Ref();
-  Register(ctx);
-  if (xTaskCreatePinnedToCore(SaveTask, "DuktapeVFSSave", 5*512, this,
-      CONFIG_OVMS_SC_JAVASCRIPT_DUKTAPE_PRIORITY-1, NULL, CORE(1)) != pdPASS)
-    {
-    Deregister(ctx);
-    Unref();
-    m_error = "cannot start task";
-    CallMethod(ctx, "fail");
-    return;
-    }
-  //ESP_LOGD(TAG, "DuktapeVFSSave('%s'): started", m_path.c_str());
-  }
-
-void DuktapeVFSSave::SaveTask(void *param)
-  {
-  DuktapeVFSSave *me = (DuktapeVFSSave*)param;
-  me->Save();
-  me->Unref();
-  vTaskDelete(NULL);
-  }
-
-void DuktapeVFSSave::Save()
-  {
-  // create path:
-  size_t n = m_path.rfind('/');
-  if (n != 0 && n != std::string::npos)
-    {
-    std::string dir = m_path.substr(0, n);
-    if (!path_exists(dir) && mkpath(dir) != 0)
-      {
-      m_error = "cannot create path";
-      RequestCallback("fail");
-      return;
-      }
-    }
-
-  // save data:
-  if (m_append)
-    m_file.open(m_path, std::ios::out | std::ios::binary | std::ios::app);
-  else
-    m_file.open(m_path, std::ios::out | std::ios::binary | std::ios::trunc);
-  if (m_file.fail())
-    {
-    m_error = "cannot open file";
-    RequestCallback("fail");
-    return;
-    }
-  m_file.write(&m_data[0], m_data.size());
-  bool wfail = m_file.fail();
-  m_file.close();
-
-  // free buffer:
-  m_data.clear();
-  m_data.shrink_to_fit();
-
-  if (wfail || m_file.fail())
-    {
-    m_error = "write error";
-    RequestCallback("fail");
-    }
-  else
-    {
-    m_error = "";
-    RequestCallback("done");
-    }
-  }
-
-duk_ret_t DuktapeVFSSave::CallMethod(duk_context *ctx, const char* method, void* data /*=NULL*/)
-  {
-  if (!ctx)
-    {
-    RequestCallback(method, data);
-    return 0;
-    }
-  OvmsRecMutexLock lock(&m_mutex);
-  if (!IsCoupled()) return 0;
-
-  duk_require_stack(ctx, 3);
-  int entry_top = duk_get_top(ctx);
-
-  bool deregister = false;
-
-  while (method)
-    {
-    const char* followup_method = NULL;
-
-    // check method:
-    int obj_idx = Push(ctx);
-    duk_get_prop_string(ctx, obj_idx, method);
-    bool callable = duk_is_callable(ctx, -1);
-    duk_pop(ctx);
-    if (callable) duk_push_string(ctx, method);
-    int nargs = 0;
-
-    // create results & method arguments:
-    if (strcmp(method, "done") == 0)
-      {
-      // done():
-      followup_method = "always";
-      //ESP_LOGD(TAG, "DuktapeVFSSave('%s'): done", m_path.c_str());
-
-      // clear request.error:
-      duk_push_string(ctx, "");
-      duk_put_prop_string(ctx, obj_idx, "error");
-      }
-    else if (strcmp(method, "fail") == 0)
-      {
-      // fail(error):
-      followup_method = "always";
-      //ESP_LOGD(TAG, "DuktapeVFSSave('%s'): failed error='%s'", m_path.c_str(), m_error.c_str());
-
-      // set request.error:
-      duk_push_string(ctx, m_error.c_str());
-      duk_dup(ctx, -1);
-      duk_put_prop_string(ctx, obj_idx, "error");
-      nargs++;
-      }
-    else if (strcmp(method, "always") == 0)
-      {
-      // always():
-      deregister = true;
-      }
-
-    // call method:
-    if (callable)
-      {
-      //ESP_LOGD(TAG, "DuktapeVFSSave: calling method '%s' nargs=%d", method, nargs);
-      if (duk_pcall_prop(ctx, obj_idx, nargs) != 0)
-        {
-        DukOvmsErrorHandler(ctx, -1);
-        }
-      }
-
-    // clear stack:
-    duk_pop_n(ctx, duk_get_top(ctx) - entry_top);
-
-    // followup call:
-    method = followup_method;
-    } // while (method)
-
-  // allow GC:
-  if (deregister) Deregister(ctx);
-  return 0;
-  }
-
-////////////////////////////////////////////////////////////////////////////////
 // DuktapeObjectRegistration
 //
 // An object to record a Duktape object registration for OVMS
@@ -1833,18 +880,6 @@ OvmsDuktape::OvmsDuktape()
   extern const char mod_json_js_start[]     asm("_binary_json_js_start");
   extern const char mod_json_js_end[]       asm("_binary_json_js_end");
   RegisterDuktapeModule(mod_json_js_start, mod_json_js_end - mod_json_js_start, "JSON");
-
-  // Register standard functions
-  RegisterDuktapeFunction(DukOvmsPrint, 1, "print");
-  RegisterDuktapeFunction(DukOvmsAssert, 2, "assert");
-  DuktapeObjectRegistration* dt_http = new DuktapeObjectRegistration("HTTP");
-  dt_http->RegisterDuktapeFunction(DuktapeHTTPRequest::Create, 1, "request"); // legacy
-  dt_http->RegisterDuktapeFunction(DuktapeHTTPRequest::Create, 1, "Request");
-  RegisterDuktapeObject(dt_http);
-  DuktapeObjectRegistration* dt_vfs = new DuktapeObjectRegistration("VFS");
-  dt_vfs->RegisterDuktapeFunction(DuktapeVFSLoad::Create, 1, "Load");
-  dt_vfs->RegisterDuktapeFunction(DuktapeVFSSave::Create, 1, "Save");
-  RegisterDuktapeObject(dt_vfs);
 
   // Notify the command system that scripts are ready
   MyCommandApp.NotifyDuktapeScriptsReady();
@@ -2006,6 +1041,38 @@ void OvmsDuktape::DuktapeRequestCallback(DuktapeObject* instance, const char* me
   dmsg.body.dt_callback.method = method;
   dmsg.body.dt_callback.data = data;
   DuktapeDispatch(&dmsg);
+  }
+
+OvmsWriter* OvmsDuktape::GetDuktapeWriter()
+  {
+  return duktapewriter;
+  }
+
+void OvmsDuktape::DukGetCallInfo(duk_context *ctx, std::string *filename, int *linenumber, std::string *function)
+  {
+  duk_require_stack(ctx, 3);
+  for (int i=-2; ; i--)
+    {
+    *filename = "";
+    *function = "";
+    duk_inspect_callstack_entry(ctx, i);
+    duk_get_prop_string(ctx, -1, "lineNumber");
+    *linenumber = duk_get_number_default(ctx, -1, 0);
+    duk_pop(ctx);
+    if (duk_get_prop_string(ctx, -1, "function"))
+      {
+      duk_get_prop_string(ctx, -1, "fileName");
+      *filename = duk_get_string_default(ctx, -1, "");
+      duk_pop(ctx);
+      duk_get_prop_string(ctx, -1, "name");
+      *function = duk_get_string_default(ctx, -1, "");
+      duk_pop(ctx);
+      }
+    duk_pop(ctx);
+    // skip internal modules:
+    if (!startsWith(*filename, "int/"))
+      break;
+    }
   }
 
 void *DukAlloc(void *udata, duk_size_t size)
